+++ Howest.Prog.Cia.CurrencySwapper.Wpf/MainWindow.xaml.cs
@@ -21,7 +21,7 @@ namespace Howest.Prog.Cia.CurrencySwapper.Wpf
     /// </summary>
     public partial class MainWindow : Window
     {
-        private const double EurToUsdRate = 1.0813771; //1 EUR is altijd 1,08 USD
+        private const decimal EurToUsdRate = 1.0813771M; //1 EUR is altijd 1,08 USD
         private AmountValidator validator;
         private CurrencyConverter converter;
 
@@ -37,12 +37,12 @@ namespace Howest.Prog.Cia.CurrencySwapper.Wpf
         {
             string userInput = txtInput.Text;
 
-            if (double.TryParse(userInput, out double amount))
+            if (decimal.TryParse(userInput, out decimal amount))
             {
                 var validationResult = validator.Validate(amount);
                 if (validationResult.IsValid)
                 {
-                    double convertedAmount = converter.Convert(amount, EurToUsdRate);
+                    decimal convertedAmount = converter.Convert(amount, EurToUsdRate);
 
                     txtOutput.Text = $"{amount} EUR = {convertedAmount:N2} USD";
                 }
